refactor: extract GitHubExportProvider+RepositoryDiscovery (#639 slice E)
Fifth slice of #639 (after #873/#878/#880/#882). Small independent
slice — the repository-discovery pair.

Moved (2 methods):
- func fetchRepositories(token:) — already public, unchanged
- makeRepositoryListRequest(token:page:) — stays private (single-caller fetchRepositories moves with)

Zero visibility widenings. `fetchRepositories` uses `session` which
was already widened by #880 (Duplicates slice).

Line-count delta:
- GitHubExportProvider.swift: 354 → 272 (-82 lines)
- GitHubExportProvider+RepositoryDiscovery.swift (new): 87 lines
- xcodeproj: +4 (xcodegen)

Test coverage: GitHubExportProviderTests passes unchanged. External
callers at ExportService.swift:21 and GitHubExportProviderTests.swift:426
continue to resolve fetchRepositories via extension.

Closes #884.
